refactor: 통합캘린더 네이밍 변경
